Create new --manual-html and --new-build-directory flags for build
Previously, every build would destroy the former build directory.

This commit introduces two new option flags for the build command.

-n or --new-build-directory causes the build script to use the previous
build directory if one already existed.

-m or --manual-html causes the build script not to create an index.html
file. This is for people who want to supply their own custom index.html,
running Canopy on an existing website. These people should either copy
their index.html into the build directory in a build script, or should
put an index.html file in the build directory and not use the
--new-build-directory option which would erase it.

+++ bin/commands/build.js
@@ -3,35 +3,37 @@ const dedent = require('dedent-js');
 const child_process = require('child_process');
 const buildProject = require('./build/build_project');
 
-function build({ symlinks, projectPathPrefix, hashUrls }) {
+function build({ symlinks, projectPathPrefix, hashUrls, newBuildDirectory, manualHtml }) {
 	if (!fs.existsSync('./topics')) throw 'There must be a topics directory present, try running "canopy init"';
 	if (!fs.existsSync('./.canopy_default_topic')) throw 'There must be a default topic dotfile present, try running "canopy init"';
 
 	let defaultTopic = fs.readFileSync('.canopy_default_topic').toString().trim();
 	canopyLocation = child_process.execSync("echo ${CANOPY_LOCATION:-$(readlink -f $(which canopy) | xargs dirname | xargs dirname)}").toString().trim();
 
-	fs.rmSync('build', { recursive: true, force: true });
-	fs.mkdirSync('build');
+	if (newBuildDirectory) fs.rmSync('build', { recursive: true, force: true });
+	fs.ensureDirSync('build');
 
-	let html = dedent`
-		<html>
-		<head>
-		<meta charset="utf-8">
-		</head>
-		<body>
-		<div
-		  id="_canopy"
-		  data-default-topic="${defaultTopic}"
-		  data-project-path-prefix="${projectPathPrefix}"
-		  data-hash-urls="${hashUrls || ''}">
-		</div>
-		<script src="${projectPathPrefix}/canopy.js"></script>
-		</body>
-		</html>`;
+	if (!manualHtml) {
+		let html = dedent`
+			<html>
+			<head>
+			<meta charset="utf-8">
+			</head>
+			<body>
+			<div
+			  id="_canopy"
+			  data-default-topic="${defaultTopic}"
+			  data-project-path-prefix="${projectPathPrefix}"
+			  data-hash-urls="${hashUrls || ''}">
+			</div>
+			<script src="${projectPathPrefix}/canopy.js"></script>
+			</body>
+			</html>`;
+
+		fs.writeFileSync('build/index.html', html);
+	}
 
 	buildProject('.', symlinks);
-
-	fs.writeFileSync('build/index.html', html);
 
 	if (symlinks) {
 		let topicDirectories = getDirectories('build');
@@ -44,10 +46,20 @@ function build({ symlinks, projectPathPrefix, hashUrls }) {
 		});
 	}
 
-	if (!fs.existsSync(`${canopyLocation}/dist/canopy.js`)) throw 'No Canopy js build found';
+	if (!fs.existsSync(`${canopyLocation}/dist/canopy.js`)) {
+		throw 'No Canopy js build found';
+	}
+
 	fs.copyFileSync(`${canopyLocation}/dist/canopy.js`, 'build/canopy.js');
-	if (fs.existsSync(`${canopyLocation}/dist/canopy.js.map`)) fs.copyFileSync(`${canopyLocation}/dist/canopy.js.map`, 'build/canopy.js.map');
-	if (fs.existsSync(`assets`)) fs.copySync('assets', 'build/_assets', { overwrite: true });
+
+	if (fs.existsSync(`${canopyLocation}/dist/canopy.js.map`)) {
+		fs.copyFileSync(`${canopyLocation}/dist/canopy.js.map`, 'build/canopy.js.map');
+	}
+
+	if (fs.existsSync(`assets`)) {
+		fs.copySync('assets', 'build/_assets', { overwrite: true });
+	}
+
 	console.log(`Built at: ${'' + new Date()}`)
 }
